% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spss_attribs.R
\name{add_spss_labels}
\alias{add_spss_labels}
\title{Add variable labels.}
\usage{
add_spss_labels(var, df, stored_attribs)
}
\arguments{
\item{var}{variable to add variable label to}

\item{df}{data.frame containing the variable}

\item{stored_attribs}{list object that is the result of calling
rsegmenter::store_spss_attr}
}
\description{
Add variable labels.
}
