% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spss_attribs.R
\name{add_spss_value_codes}
\alias{add_spss_value_codes}
\title{Add value codes.}
\usage{
add_spss_value_codes(var, df, data_map)
}
\arguments{
\item{var}{variable to add value codes to}

\item{df}{data.frame containing the variable}

\item{data_map}{data map that is the result of calling
rsegmenter::extract_datamap}
}
\description{
Add value codes.
}
