% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spss_attribs.R
\name{add_spss_value_labels}
\alias{add_spss_value_labels}
\title{Add value labels. This function should always be called after calling
rsegmenter::add_spss_value_codes}
\usage{
add_spss_value_labels(var, df, data_map)
}
\arguments{
\item{var}{variable to add value labels to}

\item{df}{data.frame containing the variable}

\item{data_map}{data map that is the result of calling
rsegmenter::extract_datamap}
}
\description{
Add value labels. This function should always be called after calling
rsegmenter::add_spss_value_codes
}
