% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_table.R
\name{add_table_labels}
\alias{add_table_labels}
\title{create row percent profile tables of all specified variables by segment variable}
\usage{
add_table_labels(prof_table, table_labels)
}
\arguments{
\item{df}{should be a dataframe of numeric variables}

\item{factor_vars}{should be a character vector of variable names}

\item{numeric_vars}{should be a character vector of variable names}

\item{...}{segment variables}
}
\description{
create row percent profile tables of all specified variables by segment variable
}
