% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_loading.R
\name{avg_loading}
\alias{avg_loading}
\title{assigns segments based on average factor loading}
\usage{
avg_loading(df, myloadings)
}
\arguments{
\item{myloadings}{factor loadings}
}
\description{
assigns segments based on average factor loading
}
