% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_ind_tab.r
\docType{data}
\name{col_ind_tab}
\alias{col_ind_tab}
\title{col_ind_tab.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 7 columns.
}
\usage{
col_ind_tab
}
\description{
col_ind_tab.
}
\keyword{datasets}
