% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_tab.r
\docType{data}
\name{col_tab}
\alias{col_tab}
\title{col_tab.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40 rows and 8 columns.
}
\usage{
col_tab
}
\description{
col_tab.
}
\keyword{datasets}
