% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{creates confusion matrix}
\usage{
confusion_matrix(df, actual, predicted, weight_var = NULL)
}
\arguments{
\item{df}{data.frame of numeric variables}

\item{actual}{ground truth data}

\item{predicted}{predicted data}

\item{weight_var}{if not NULL, a vector that contains weights for each observation. The NULL
case is equivalent to all cases being weighted 1.}
}
\description{
creates confusion matrix
}
