% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_algorithm.R
\name{create_algorithm}
\alias{create_algorithm}
\title{Creates an excel workbook with two worksheets. The first sheet contains the classification
function coefficients from a model created using rsegmenter::lda. The second sheet contains the training data
and excel formulae to assign the training data to a segment, using the classification function
coefficients.}
\usage{
create_algorithm(df, id, indeps, ldamodel)
}
\arguments{
\item{df}{data.frame of numeric variables.}

\item{id}{unique row identifier.}

\item{indeps}{predictor variables.}

\item{ldamodel}{lda model object.}
}
\description{
Creates an excel workbook with two worksheets. The first sheet contains the classification
function coefficients from a model created using rsegmenter::lda. The second sheet contains the training data
and excel formulae to assign the training data to a segment, using the classification function
coefficients.
}
