% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab_segmentation.R
\name{crosstab_segmentation}
\alias{crosstab_segmentation}
\title{create a segment variable by assigning a unique code to every combination of the input variables. input variables are sorting in
descending order, from the left most column to the right most column}
\usage{
crosstab_segmentation(df, ...)
}
\arguments{
\item{df}{data.frame of numeric variables}

\item{...}{variables to be crosstabbed}
}
\description{
create a segment variable by assigning a unique code to every combination of the input variables. input variables are sorting in
descending order, from the left most column to the right most column
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1),myweight=c(1,2,1))
crosstab_segmentation(df = mydf,col1,col2,col3)
}
