% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_missings.R
\name{explore_missings}
\alias{explore_missings}
\title{Explore missing values at variable and respondent levels}
\usage{
explore_missings(df, vars)
}
\arguments{
\item{df}{data.frame of numeric variables.}

\item{vars}{variables to be explore missings for.}
}
\description{
Explore missing values at variable and respondent levels
}
