% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_table.R
\name{export_factor_loadings}
\alias{export_factor_loadings}
\title{Exports factor loadings to .xlsx format. Each table will be output on a separate worksheet.
The max loading per row is highlighted.}
\usage{
export_factor_loadings(fac_loadings, filename)
}
\arguments{
\item{fac_loadings}{output from rsegmenter::extract_factor_loadings.}

\item{filename}{file name to use when saving.}
}
\description{
Exports factor loadings to .xlsx format. Each table will be output on a separate worksheet.
The max loading per row is highlighted.
}
\examples{

}
