% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_table.R
\name{export_profile_tables}
\alias{export_profile_tables}
\title{Exports cross tables to .xlsx format.}
\usage{
export_profile_tables(prof_table, min_index = 80, max_index = 120, filename)
}
\arguments{
\item{prof_table}{should be the output from rsegmenter::profile_table}

\item{min_index}{numeric value for highlighting under-indexing}

\item{max_index}{numeric value for highlighting over-indexing}

\item{filename}{character string naming a file}
}
\description{
Exports cross tables to .xlsx format.
}
\examples{
df <- data.frame(var1=c(1,2,3),var2=c(2,1,3),var3=c(3,2,1))
segments <- factor_segmentation(df,c(2:3))
profile_table <- profile_table(df,factor_vars = c("var1","var2","var3"),
numeric_vars = NULL,segment_vars,table_labels)
export_profile_tables(profile_table,min_index=80,max_index=120,"profile_tables.xlsx")
}
