% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datamap.R
\name{extract_datamap}
\alias{extract_datamap}
\title{Returns a data.frame containing a mapping between variables, value codes and value labels.}
\usage{
extract_datamap(df)
}
\arguments{
\item{df}{data.frame of numeric variables. The file must have been imported using one of haven::read_sav or
rsegmenter::read_data. See the test_datamap example that comes with the package.}
}
\description{
Returns a data.frame containing a mapping between variables, value codes and value labels.
}
