% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_segments.R
\name{extract_factor_scores}
\alias{extract_factor_scores}
\title{Extract factor scores for each segment solution from factor analysis stores each set of scores as a data.frame as element of a list.}
\usage{
extract_factor_scores(seglist)
}
\arguments{
\item{seglist}{an output from the rsegmenter::factor_segmentation() function. See more details about factor_segmentation()
in the package documentation.}
}
\description{
Extract factor scores for each segment solution from factor analysis stores each set of scores as a data.frame as element of a list.
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1))
segment_solutions <- factor_segmentation(mydf,vars = c("col1","col2","col3"),
weight_var = NULL, num_sols = c(2:3), rotate = "varimax",
scores = TRUE, fac_assign = "avg_loading")
extract_factor_scores(segment_solutions)
}
