% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_segments.R
\name{extract_kmeans_segments}
\alias{extract_kmeans_segments}
\title{extract segments from kmeans segment solutions and stores in a dataframe}
\usage{
extract_kmeans_segments(df)

extract_kmeans_segments(df)
}
\arguments{
\item{df}{should be an output from the kantarsegmentr::segmentation() function}

\item{df}{should be an output from the kantarsegmentr::segmentation() function}
}
\description{
extract segments from kmeans segment solutions and stores in a dataframe

extract segments from kmediods segment solutions and stores in a dataframe
}
