% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_segments.R
\name{extract_lca_segments}
\alias{extract_lca_segments}
\title{extract segments from lca segment solutions and stores in a dataframe}
\usage{
extract_lca_segments(seglist)
}
\arguments{
\item{seglist}{must be an output from the kantarsegmentr::segmentation() function}
}
\description{
extract segments from lca segment solutions and stores in a dataframe
}
