% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_segmentation.R
\name{factor_segmentation}
\alias{factor_segmentation}
\title{runs factor analysis with varimax rotation using the psych package}
\usage{
factor_segmentation(df, vars, num_sols, weight_var)
}
\arguments{
\item{df}{must be a data.frame of numeric variables}

\item{vars}{must be a string of variable names to operate on.}

\item{num_sols}{should be a numeric vector specifying the minimum and maximum number of factors to extract}
}
\description{
runs factor analysis with varimax rotation using the psych package
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1))
factor_segmentation(df = mydf, vars = c("col1","col2","col3"),num_sols=c(3,5))
}
