% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_segmentation.R
\name{hierarchical_segmentation}
\alias{hierarchical_segmentation}
\title{runs hierarchical clustering using hclust}
\usage{
hierarchical_segmentation(
  df,
  vars,
  dist_metric = "euclidean",
  linkage = "ward.D2"
)
}
\arguments{
\item{df}{data.frame of numeric variables}

\item{vars}{character vector of variable names}

\item{dist_metric}{name of distance metric to use. see stats::dist for more distance options}

\item{linkage}{name of linkage method to use. see stats::hclust for more linkage options}
}
\description{
runs hierarchical clustering using hclust
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1))
hierarchical_segmentation(df = mydf, vars = c("col1","col2","col3"))
}
