% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmediods_segmentation.R
\name{kmediods_segmentation}
\alias{kmediods_segmentation}
\title{runs kmediods}
\usage{
kmediods_segmentation(dist_mat, num_sols)
}
\arguments{
\item{dist_mat}{should be a distance matrix}

\item{num_sols}{should be a numeric vector specifying the minimum and maximum number of segments to extract}
}
\description{
runs kmediods
}
