% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca_segmentation.R
\name{lca_segmentation}
\alias{lca_segmentation}
\title{runs lca using the polca package}
\usage{
lca_segmentation(df, vars, num_sols)
}
\arguments{
\item{df}{must be a data.frame of numeric variables}

\item{vars}{must be a string of variable names to operate on.
These variables must be numeric}

\item{num_sols}{should be a numeric vector specifying the minimum and maximum number of factors to extract}
}
\description{
runs lca using the polca package
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1))
lca_segmentation(df = mydf, vars = c("col1","col2","col3"),num_sols=c(3,5))
}
