% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{lda}
\alias{lda}
\title{Runs lda from the MASS package}
\usage{
lda(df, dep, indeps, prior)
}
\arguments{
\item{df}{must be a data.frame}

\item{dep}{must be a string of the dependent variable name}

\item{indeps}{character vector of predictor variable names}

\item{prior}{must be a numeric vector of prior probabilities.
length(prior) must equal number of unique values in the dependent variable}
}
\description{
Returns the lda model object from the MASS package with a additional
elements that holds the fishers classification function coefficients and the
predicted outcome
}
\examples{
df <- rsegmenter::test_seg_unlabelled

lda(df, dep="seg1", indeps=c("seg2","seg3","seg4"),prior=rep(1/4,4))

}
