% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis.R
\name{mahalanobis_outlier}
\alias{mahalanobis_outlier}
\title{Determine outliers based on mahalanobis distance}
\usage{
mahalanobis_outlier(df, vars, pval_cutoff = 0.001)
}
\arguments{
\item{df}{data.frame of numeric variables.}

\item{vars}{variables to be used in the outlier calculation.}

\item{pval_cutoff}{cutoff to classify an outlier. Any pvalue less than the cutoff
will be classified as an outlier}
}
\description{
Determine outliers based on mahalanobis distance
}
