% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_table.R
\name{profile_table}
\alias{profile_table}
\title{Creates a data.frame of 4 sets of cross tables. These are raw count, col \%, row \% and col \% index tables}
\usage{
profile_table(
  df,
  factor_vars = NULL,
  numeric_vars = NULL,
  weight_var,
  segment_vars,
  table_labels
)
}
\arguments{
\item{df}{data.frame of input variables}

\item{factor_vars}{character vector of variable names that are to be treated as factor.
Factor variables will have counts shown for each level of each variable.}

\item{numeric_vars}{character vector of variable names that are to be treated as numeric.
Numeric variables will have means shown for variable.}

\item{weight_var}{name of the variable holding case/row weights.
If data is un-weighted, specify a vector of 1's with length equal to number of rows of df}

\item{segment_vars}{character vector segment variables to profile by}
}
\description{
Creates a data.frame of 4 sets of cross tables. These are raw count, col \%, row \% and col \% index tables
}
