% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_table.R
\name{profile_table_col_perc}
\alias{profile_table_col_perc}
\title{create column percent profile tables of all specified variables by segment variable}
\usage{
profile_table_col_perc(
  df,
  factor_vars = NULL,
  numeric_vars = NULL,
  weight_var,
  segment_var
)
}
\arguments{
\item{df}{should be a dataframe of numeric variables}

\item{factor_vars}{should be a character vector of variable names}

\item{numeric_vars}{should be a character vector of variable names}

\item{segment_var}{should be a character vector of segment variable name}
}
\description{
create column percent profile tables of all specified variables by segment variable
}
