% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_table.R
\name{profile_table_raw}
\alias{profile_table_raw}
\title{Create raw count profile tables of all specified variables by banner variable.}
\usage{
profile_table_raw(
  df,
  factor_vars = NULL,
  numeric_vars = NULL,
  weight_var = NULL,
  banner_var
)
}
\arguments{
\item{df}{data.frame of input variables}

\item{factor_vars}{variables that are to be treated as factors in the table output.
Will produce counts for each value for each variable. If weight_var is provided, will
produce weighted counts.}

\item{numeric_vars}{variables that are to be treated as numeric in the table output.
Will produce means for each variable. if weight_var is provided, will produce weighted means.}

\item{weight_var}{if not NULL, a vector that contains weights for each observation. The NULL
case is equivalent to all cases being weighted 1.}

\item{banner_var}{name of the variable to appear along the top of the table.}
}
\description{
Create raw count profile tables of all specified variables by banner variable.
}
