% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_table.R
\name{profile_table_totals}
\alias{profile_table_totals}
\title{Create total summary tables.}
\usage{
profile_table_totals(df, weight_var = NULL, banner_var)
}
\arguments{
\item{df}{data.frame of input variables}

\item{weight_var}{if not NULL, a vector that contains weights for each observation. The NULL
case is equivalent to all cases being weighted 1.}

\item{banner_var}{name of the variable to appear along the top of the table.}
}
\description{
Create total summary tables.
}
