% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_tab.r
\docType{data}
\name{raw_tab}
\alias{raw_tab}
\title{raw_tab.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 8 columns.
}
\usage{
raw_tab
}
\description{
raw_tab.
}
\keyword{datasets}
