% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{loads data from various sources
for now it is limited to .csv and .sav
uses vroom for fast loading of csv
uses haven for sav files}
\usage{
read_data(infile)
}
\arguments{
\item{infile}{should be the name of the file to be read in}
}
\description{
loads data from various sources
for now it is limited to .csv and .sav
uses vroom for fast loading of csv
uses haven for sav files
}
