% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_tab.r
\docType{data}
\name{row_tab}
\alias{row_tab}
\title{row_tab.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40 rows and 7 columns.
}
\usage{
row_tab
}
\description{
row_tab.
}
\keyword{datasets}
