% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation}
\alias{segmentation}
\title{runs kmeans, factor analsysis, kmediods}
\usage{
segmentation(df, min_segs = 3, max_segs = 5, weight_var)
}
\arguments{
\item{df}{should be a dataframe of numeric variables}

\item{min_segs}{should be an integer specifying the minimum number of segments to extract, defaults to 3}

\item{max_segs}{should be an integer specifying the maximum number of segments to extract, defaults to 5}
}
\description{
runs kmeans, factor analsysis, kmediods
}
