% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_spss_attribs.R
\name{store_spss_attr}
\alias{store_spss_attr}
\title{Returns a list containing the label, format and display width attribute information
for all applicable variables}
\usage{
store_spss_attr(df)
}
\arguments{
\item{df}{data.frame.}
}
\description{
Returns a list containing the label, format and display width attribute information
for all applicable variables
}
