% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datamap.r
\docType{data}
\name{test_datamap}
\alias{test_datamap}
\title{datamap corresponding to test_seg_labelled data file.}
\format{
An object of class \code{data.frame} with 48 rows and 5 columns.
}
\usage{
test_datamap
}
\description{
datamap corresponding to test_seg_labelled data file.
}
\keyword{datasets}
