% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_seg_unlabelled.r
\docType{data}
\name{test_seg_unlabelled}
\alias{test_seg_unlabelled}
\title{Randomly generated data in SPSS format with 1000 rows, 2 demographic variables, respondent weight variable
and 10 segmentation input variables. Variables are unlabelled.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1000 rows and 14 columns.
}
\usage{
test_seg_unlabelled
}
\description{
Randomly generated data in SPSS format with 1000 rows, 2 demographic variables, respondent weight variable
and 10 segmentation input variables. Variables are unlabelled.
}
\keyword{datasets}
