% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_two_box.R
\name{top_two_box}
\alias{top_two_box}
\title{Creates a 0,1 coded variable. The two highest values become a 1 and the rest becomes a 0}
\usage{
top_two_box(df, vars, impute_type = "none")
}
\arguments{
\item{df}{must be a data.frame}

\item{vars}{must be a string of variable names to operate on.
These variables must be numeric}
}
\description{
Creates a 0,1 coded variable. The two highest values become a 1 and the rest becomes a 0
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1))
top_box(df = mydf, vars = c("col1","col2","col3"))
}
