% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_table.R
\name{add_table_labels}
\alias{add_table_labels}
\title{Adds user friendly labels to the profile tables}
\usage{
add_table_labels(prof_table, table_labels)
}
\arguments{
\item{prof_table}{should be a dataframe of numeric variables}

\item{table_labels}{should be a character vector of variable names}
}
\description{
Adds user friendly labels to the profile tables
}
