% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_flatline.R
\name{check_flatline}
\alias{check_flatline}
\title{Returns the max percentage of responses that are the same per row.}
\usage{
check_flatline(myvar)
}
\arguments{
\item{myvar}{must be a numeric vector to operate on.}
}
\description{
Returns the max percentage of responses that are the same per row.
The closer this return value is to 1, the more serious the flatline problem
is for that row.
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1))
check_flatline(mydf$col1)
}
