% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_flatline.R
\name{check_flatline}
\alias{check_flatline}
\title{gets the number of responses that are the same per person}
\usage{
check_flatline(df)
}
\arguments{
\item{df}{should be a dataframe of numeric variables}
}
\description{
gets the number of responses that are the same per person
}
