% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_flatline.R
\name{check_flatline}
\alias{check_flatline}
\title{check_flatline returns the max percentage of responses that are the same per row.}
\usage{
check_flatline(df, vars, impute_type = "none")
}
\arguments{
\item{df}{must be a data.frame}

\item{vars}{must be a string of variable names to operate on.
These variables must be numeric}

\item{impute_type}{must be a string of one of "none","mode","mean","min","max"}
}
\description{
The closer this return value is to 1, the more serious the flatline problem is for that row.
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1))
check_flatline(df = mydf, vars = c("col1","col2","col3"), impute_type = "none")
}
