% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{creates confusion matrix}
\usage{
confusion_matrix(df, actual, predicted, weight_var)
}
\arguments{
\item{df}{data.frame of numeric variables}

\item{actual}{character vector of variable name of ground truth}

\item{predicted}{character vector of variable name of predicted}

\item{weight_var}{name of the variable holding case/row weights.}
}
\description{
creates confusion matrix
}
\examples{
mydf <- data.frame(actual=c(1,2,3),predicted=c(1,3,2),weight=c(1,1,1))
confusion_matrix(df = mydf,actual,predicted,weight)
}
