% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_table.R
\name{export_profile_tables}
\alias{export_profile_tables}
\title{create row percent profile tables of all specified variables by segment variable}
\usage{
export_profile_tables(prof_table, min_index = 80, max_index = 120,
  filename)
}
\arguments{
\item{df}{should be a dataframe of numeric variables}

\item{factor_vars}{should be a character vector of variable names}

\item{numeric_vars}{should be a character vector of variable names}

\item{...}{segment variables}
}
\description{
create row percent profile tables of all specified variables by segment variable
}
