% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_table.R
\name{export_profile_tables}
\alias{export_profile_tables}
\title{Exports cross tables to .xlsx format. Each table will be output on a separate worksheet.}
\usage{
export_profile_tables(prof_table, min_index = 80, max_index = 120, filename)
}
\arguments{
\item{prof_table}{output from rsegmenter::profile_table.}

\item{min_index}{numeric value for highlighting under-indexing.}

\item{max_index}{numeric value for highlighting over-indexing.}

\item{filename}{file name to use when saving.}
}
\description{
Exports cross tables to .xlsx format. Each table will be output on a separate worksheet.
}
\examples{
df <- rsegmenter::test_seg_unlabelled

table_labels <- rsegmenter::extract_datamap(rsegmenter::test_seg_labelled)

prof_table <- profile_table(df, category_vars = c("demog1","demog2"), numeric_vars = NULL,
weight_var = "weight", banner_vars = "seg1", table_labels = table_labels)

export_profile_tables(prof_table = prof_table, filename="profile_table.xlsx")

}
