% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datamap.R
\name{extract_datamap}
\alias{extract_datamap}
\title{extract_datamap returns a dataframe containing a mapping between variables, value codes and value labels.}
\usage{
extract_datamap(df)
}
\arguments{
\item{df}{must be a .sav file stored as data.frame.
Importing the .sav file using haven::read_sav or
rsegmenter::read_data is the preferred methods}
}
\description{
extract_datamap returns a dataframe containing a mapping between variables, value codes and value labels.
}
