% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datamap.R
\name{extract_datamap}
\alias{extract_datamap}
\title{extract_datamap returns a dataframe containing a mapping between variables, value codes and value labels.}
\usage{
extract_datamap(df)
}
\arguments{
\item{df}{must be a data.frame that is read in using haven::read_sav}
}
\description{
extract_datamap returns a dataframe containing a mapping between variables, value codes and value labels.
}
