% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_segments.R
\name{extract_factor_segments}
\alias{extract_factor_segments}
\title{extract segments from factor analysis segment solutions and stores in a dataframe}
\usage{
extract_factor_segments(seglist)
}
\arguments{
\item{seglist}{must be an output from the rsegmenter::segmentation() function}
}
\description{
extract segments from factor analysis segment solutions and stores in a dataframe
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1))
segment_solutions <- segmentation(mydf,c("Col1","Col2"))
extract_factor_segments(segment_solutions)
}
