% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_segments.R
\name{extract_factor_segments}
\alias{extract_factor_segments}
\title{extract segments from factor analysis segment solutions and stores in a dataframe}
\usage{
extract_factor_segments(seglist)
}
\arguments{
\item{seglist}{must be an output from the kantarsegmentr::segmentation() function}
}
\description{
extract segments from factor analysis segment solutions and stores in a dataframe
}
