% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_segments.R
\name{extract_factor_segments}
\alias{extract_factor_segments}
\title{Extract segments from factor analysis segment solutions and stores each segment as a variable in a data.frame.}
\usage{
extract_factor_segments(seglist, suffix = NULL)
}
\arguments{
\item{seglist}{an output from the rsegmenter::factor_segmentation() function. See more details about factor_segmentation()
in the package documentation.}
}
\description{
Extract segments from factor analysis segment solutions and stores each segment as a variable in a data.frame.
}
