% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_segments.R
\name{extract_kmeans_segments}
\alias{extract_kmeans_segments}
\title{Extract segments from lca segment solutions and stores each segment as a variable in a data.frame.}
\usage{
extract_kmeans_segments(seglist)
}
\arguments{
\item{seglist}{an output from the rsegmenter::lca_segmentation() function. See more details about lca_segmentation()
in the package documentation.}
}
\description{
Extract segments from lca segment solutions and stores each segment as a variable in a data.frame.
}
