% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_segments.R
\name{extract_lca_segments}
\alias{extract_lca_segments}
\title{Extract segments from lca segment solutions and stores each segment as a variable in a data.frame.}
\usage{
extract_lca_segments(seglist)
}
\arguments{
\item{seglist}{an output from the rsegmenter::lca_segmentation() function. See more details about lca_segmentation()
in the package documentation.}
}
\description{
Extract segments from lca segment solutions and stores each segment as a variable in a data.frame.
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1))
segment_solutions <- lca_segmentation(mydf,vars = c("col1","col2","col3"),
num_sols = c(2:3))
extract_lca_segments(segment_solutions)
}
