% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_segments.R
\name{extract_rotated_components}
\alias{extract_rotated_components}
\title{extract rotated components from factor analysis}
\usage{
extract_rotated_components(seglist, var_labels)
}
\arguments{
\item{seglist}{must be an output from the kantarsegmentr::segmentation() function}

\item{var_labels}{must be a data.frame.
First column must be names of the variables run in the factor analysis.
Second column must be the corresponding label}
}
\description{
extract rotated components from factor analysis
}
