% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_segmentation.R
\name{factor_segmentation}
\alias{factor_segmentation}
\title{runs factor analysis with varimax rotation using the psych package}
\usage{
factor_segmentation(df, num_sols, weight_var = NULL)
}
\arguments{
\item{df}{should be a dataframe of numeric variables}

\item{num_sols}{should be a numeric vector specifying the minimum and maximum number of factors to extract}
}
\description{
runs factor analysis with varimax rotation using the psych package
}
