% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_segmentation.R
\name{factor_segmentation}
\alias{factor_segmentation}
\title{runs factor analysis with varimax rotation using the psych package}
\usage{
factor_segmentation(
  df,
  vars,
  weight_var = NULL,
  num_sols,
  rotate = "varimax",
  scores = FALSE,
  fac_assign = "avg_loading"
)
}
\arguments{
\item{df}{data.frame of numeric variables}

\item{vars}{character vector of variable names}

\item{weight_var}{numeric vector of row weights}

\item{num_sols}{numeric vector specifying the minimum and maximum number of factors to extract}

\item{rotate}{method of rotation for factor analysis. Options are the same as psych::principal()}

\item{scores}{TRUE/FALSE to include scores in the output or not. if fac_assign="max_score", scores must be set to TRUE}

\item{fac_assign}{method to use to assign segments to rows. options are one of c("avg_loading","max_score")}
}
\description{
runs factor analysis with varimax rotation using the psych package
}
\examples{
df <- rsegmenter::test_seg_unlabelled
segment_input_vars <- c("seg1","seg2","seg3","seg4","seg5","seg6","seg7","seg8","seg9","seg10")
factor_segmentation(df = df, vars = segment_input_vars, weight_var = "weight", num_sols=c(2,3),
rotate="varimax", scores=FALSE, fac_assign="avg_loading")

}
