% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_segmentation.R
\name{factor_segmentation}
\alias{factor_segmentation}
\title{runs factor analysis with varimax rotation using the psych package}
\usage{
factor_segmentation(df, ..., num_sols, weight_var)
}
\arguments{
\item{df}{data.frame of numeric variables}

\item{...}{variables for the factor analysis}

\item{num_sols}{numeric vector specifying the minimum and maximum number of factors to extract}

\item{weight_var}{numeric vector of row weights}
}
\description{
runs factor analysis with varimax rotation using the psych package
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1),myweight=c(1,2,1))
factor_segmentation(df = mydf,col1,col2,col3,num_sols=c(3,5),myweight)
}
