% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_segmentation.R
\name{kmeans_segmentation}
\alias{kmeans_segmentation}
\title{runs kmeans}
\usage{
kmeans_segmentation(dist_mat, num_sols)
}
\arguments{
\item{dist_mat}{should be a distance matrix}

\item{num_sols}{should be a numeric vector specifying the minimum and maximum number of factors to extract}
}
\description{
runs kmeans
}
