% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{lda}
\alias{lda}
\title{Runs Linear Discriminant Analysis using the MASS package.
Returns the lda model object from the MASS package with an additional
elements that holds the fishers classification function coefficients and the
predicted outcome.}
\usage{
lda(df, id, dep, indeps, prior = NULL, create_algorithm = FALSE)
}
\arguments{
\item{df}{data.frame of numeric variables.}

\item{id}{unique row identifier.}

\item{dep}{target variable to be predicted.}

\item{indeps}{predictor variables.}

\item{prior}{numeric vector of prior probabilities. If NULL, will default
to 1 divided by number of distinct categories in the dependent variable.}

\item{create_algorithm}{TRUE/FALSE. If TRUE, will produce an excel file containing the prediction
algorithm.}
}
\description{
Runs Linear Discriminant Analysis using the MASS package.
Returns the lda model object from the MASS package with an additional
elements that holds the fishers classification function coefficients and the
predicted outcome.
}
\examples{
df <- rsegmenter::test_seg_unlabelled
lda(df, dep="seg1", indeps=c("seg2","seg3","seg4"),prior=rep(1/4,4))

}
