% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{lda}
\alias{lda}
\title{Runs lda from the MASS package}
\usage{
lda(df, dep, indeps, prior)
}
\arguments{
\item{df}{must be a data.frame}

\item{dep}{must be a string of the dependent variable name}

\item{indeps}{must be a string of variable names to operate on.
These variables must be numeric}

\item{prior}{must be a numeric vector of prior probabilities.
length(prior) must equal number of unique values in the dependent variable}
}
\description{
Returns the lda model object from the MASS package with an additional
element that holds the fishers classification function coefficients
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1))
lda(df = mydf, dep = "col1", indeps = c("col2","col3"), prior = rep(1/3,3))
}
