% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_score.R
\name{max_score}
\alias{max_score}
\title{assigns segments based on max factor score}
\usage{
max_score(myfacscores)
}
\arguments{
\item{myfacscores}{factor scores}
}
\description{
assigns segments based on max factor score
}
