% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_table.R
\name{profile_table}
\alias{profile_table}
\title{create profile raw count, col \%, row \%, col \% index tables of all specified variables by segment variable}
\usage{
profile_table(
  df,
  factor_vars = NULL,
  numeric_vars = NULL,
  weight_var,
  segment_vars,
  table_labels
)
}
\arguments{
\item{df}{must be a dataframe of numeric variables}

\item{factor_vars}{must be a character vector of variable names}

\item{numeric_vars}{must be a character vector of variable names}

\item{weight_var}{must be a numeric vector of row weights. if data is unweighted, specify a vector of 1's with length equal to number of rows of df}

\item{...}{segment variables}
}
\description{
create profile raw count, col \%, row \%, col \% index tables of all specified variables by segment variable
}
