% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_table.R
\name{profile_table_raw}
\alias{profile_table_raw}
\title{create raw count profile tables of all specified variables by segment variable}
\usage{
profile_table_raw(
  df,
  factor_vars = NULL,
  numeric_vars = NULL,
  weight_var,
  segment_var
)
}
\arguments{
\item{df}{should be a dataframe of numeric variables}

\item{factor_vars}{a character vector of variable names}

\item{numeric_vars}{a character vector of variable names}

\item{weight_var}{name of weight variable}

\item{segment_var}{a character vector of segment variable name}
}
\description{
create raw count profile tables of all specified variables by segment variable
}
