% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_scale.R
\name{reverse_scale}
\alias{reverse_scale}
\title{reverses the coding of the numeric variables provided.}
\usage{
reverse_scale(df)
}
\arguments{
\item{df}{should be a dataframe of numeric variables}
}
\description{
reverses the coding of the numeric variables provided.
}
