% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_scale.R
\name{reverse_scale}
\alias{reverse_scale}
\title{Reverses the coding of each input variable.}
\usage{
reverse_scale(myvar)
}
\arguments{
\item{myvar}{must be a numeric vector to operate on.}
}
\description{
Reverses the coding of each input variable.
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1))
reverse_scale(mydf$col1)
}
