% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation}
\alias{segmentation}
\title{runs factor analsysis and lca}
\usage{
segmentation(
  df,
  vars,
  min_segs = 3,
  max_segs = 5,
  weight_var = NULL,
  method = c("factor")
)
}
\arguments{
\item{df}{should be a data.frame of numeric variables}

\item{vars}{must be a string of variable names to operate on.
These variables must be numeric}

\item{min_segs}{must be an integer specifying the minimum number of segments to extract, defaults to 3}

\item{max_segs}{must be an integer specifying the maximum number of segments to extract, defaults to 5}

\item{method}{must be a character vector of at least one of 'factor' or 'lca'}

\item{weight}{must be a character vector of the name of the row weight or NULL}
}
\description{
runs factor analsysis and lca
}
\examples{
mydf <- data.frame(col1=c(1,2,3),col2=c(1,3,2),col3=c(1,2,1))
segmentation(df = mydf, vars = c('col1','col2','col3'),min_segs = 3, max_segs = 5,method = c('factor'))
}
