% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_box.R
\name{top_box}
\alias{top_box}
\title{creates a 0,1 coded variable. The highest value become a 1 and the rest becomes a 0}
\usage{
top_box(df)
}
\arguments{
\item{df}{should be a dataframe of numeric variables}
}
\description{
creates a 0,1 coded variable. The highest value become a 1 and the rest becomes a 0
}
